Add non-mut EventReader iter method
The part of my code that require access to the events only has a
non-mutable reference to it, even if the system itself has a mutable
handle on it. I need to dissociate the mutable part from the non-mutable
part of the EventReader API to be able to use it.